requirements.txt
kafka-python==2.0.2
psutil==5.9.5
matplotlib==3.7.1
pandas==2.0.3
numpy==1.24.3
requests==2.31.0